update the invitation routes
