+++ app/javascript/controllers/delete_evaluator_modal_controller.js
@@ -43,16 +43,16 @@ export default class extends Controller {
   deleteEvaluator(forceDelete = false) {
     const csrfToken = document.querySelector('meta[name="csrf-token"]').content
   
-    fetch(`/phases/${this.phaseIdValue}/manage_evaluators/${this.evaluatorIdValue}`, {
+    fetch(`/phases/${this.phaseIdValue}/evaluators/${this.evaluatorIdValue}`, {
       method: 'DELETE',
       headers: { 'Content-Type': 'application/json', 'X-CSRF-Token': csrfToken },
       body: JSON.stringify({ evaluator_type: this.evaluatorTypeValue, phase_id: this.phaseIdValue, force_delete: forceDelete })
     })
     .then(response => {
-      if (!response.ok) {
-        throw new Error('Network response was not ok')
-      }
-      const data = response.json()
+      if (!response.ok) { throw new Error('Network response was not ok') }
+      return response.json()
+    })
+    .then(data => {
       if (data.success) {
         this.close()
         window.location.reload()
@@ -60,13 +60,12 @@ export default class extends Controller {
         throw new Error(data.message || 'Failed to remove evaluator')
       }
     })
-    .catch(error => {
-      alert(error.message || 'An error occurred while removing the evaluator')
-    })
+    .catch(error => { alert(error.message || 'An error occurred while removing the evaluator') })
   }
 
   resetModal() {
-    this.modalDescriptionTarget.textContent = 'Deleting an evaluator from the challenge will remove the evaluator from any submissions of this challenge that the evaluator is assigned to. It will also delete any of their completed or in progress evaluations for those submissions.'
+    this.modalDescriptionTarget.textContent = 'Deleting an evaluator from the challenge will remove the evaluator from any submissions of this ' +
+     'challenge that the evaluator is assigned to. It will also delete any of their completed or in progress evaluations for those submissions.'
     this.confirmButtonTarget.textContent = 'Yes'
   }
 }
