+++ app/views/evaluators/_delete_evaluator_modal.html.erb
@@ -34,15 +34,5 @@
         </ul>
       </div>
     </div>
-    <button
-      type="button"
-      class="usa-button usa-modal__close"
-      aria-label="Close this window"
-      data-action="click->delete-evaluator-modal#close"
-    >
-      <svg class="usa-icon" aria-hidden="true" focusable="false" role="img">
-        <use xlink:href="/assets/img/sprite.svg#close"></use>
-      </svg>
-    </button>
   </div>
 </dialog>
